requirements.txt
pillow
customtkinter
pyinstaller